+++ .eslintrc.js
@@ -1,9 +1,14 @@
 module.exports = {
+  'root': true,
+  'parser': '@typescript-eslint/parser',
   'env': {
     'browser': true,
     'commonjs': true,
     'es2021': true,
   },
+  'plugins': [
+    '@typescript-eslint',
+  ],
   'extends': [
     'google',
   ],
