+++ package.json
@@ -3,8 +3,8 @@
   "version": "1.0.2",
   "main": "index.js",
   "scripts": {
-    "lint": "eslint .",
-    "package": "ncc build index.js -o dist --source-map --license licenses.txt"
+    "lint": "eslint . --ext .ts",
+    "package": "tsc && ncc build index.js -o dist --source-map --license licenses.txt"
   },
   "keywords": [],
   "author": "Ably",
@@ -13,16 +13,20 @@
   "dependencies": {
     "@actions/core": "^1.6.0",
     "@actions/github": "^5.0.0",
-    "glob": "^7.2.0",
-    "request": "^2.88.2"
+    "@actions/glob": "^0.3.0",
+    "got": "^12.1.0"
   },
   "devDependencies": {
+    "@types/node": "^18.11.14",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
     "@vercel/ncc": "^0.34.0",
     "eslint": "^8.17.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-n": "^15.2.2",
-    "eslint-plugin-promise": "^6.0.0"
+    "eslint-plugin-promise": "^6.0.0",
+    "typescript": "^4.7.4"
   }
 }
